Did a MAJOR overhaul to allow the user to define his own higher-order functions. All types are now unsigned 64 bit integers. which could function either as values or as pointers. The closure value is a int64 pointer to a struct of 2 int64 pointers to the function and the env struct

+++ c-src/constants.c
@@ -1,6 +1,6 @@
-const int TRUE_VALUE = 0x6F;
-const int FALSE_VALUE = 0x2F;
-const int NIL_VALUE = 0x3F;
+const unsigned long TRUE_VALUE = 0x6F;
+const unsigned long FALSE_VALUE = 0x2F;
+const unsigned long NIL_VALUE = 0x3F;
 
 const int FIXNUM_TAG = 0;
 const int FIXNUM_TAG_LEN = 2;
